scheduler_task/requirements.txt
Flask
azure-storage-queue
python-dotenv
azure-identity
fastapi
uvicorn[standard]
pydantic
azure-identity